+++ BxInputController/Sources/Rows/Variants/View/BxInputSegmentedVariantCell.swift
@@ -18,12 +18,14 @@ open class BxInputSegmentedVariantCell : UITableViewCell, BxInputTitleCell {
     
     /// caption of row
     @IBOutlet weak open var titleLabel: UILabel!
+    @IBOutlet var titleLeading: NSLayoutConstraint!
     /// caption of title for the row. This is only option field
     @IBOutlet public weak var subtitleLabel: UILabel?
     /// for managment of position subtitle
     @IBOutlet public weak var subtitleBottomMargin: NSLayoutConstraint?
     /// rate control
     @IBOutlet weak open var segmentedControl: UISegmentedControl!
+    @IBOutlet public var segmentedLeading: NSLayoutConstraint!
     
     public weak var delegate: BxInputSegmentedVariantDelegate? = nil